1.update

+++ event/event_room_join_fail_by_full.go
@@ -10,6 +10,6 @@ func RoomJoinFailByFull(ctx *EContext, data string) {
 	dataMap := make(map[string]interface{})
 	_ = json.Unmarshal([]byte(data), &dataMap)
 
-	command.PrintNotice("Join room failed. Room " + strconv.Itoa(dataMap["roomId"].(int)) + " player count is full!")
+	command.PrintNotice("Join room failed. Room " + strconv.Itoa(int(dataMap["roomId"].(float64))) + " player count is full!")
 	ShowOptions(ctx, data)
 }
